+++ Warehouse.Wpf.Models/ProductPriceUpdate.cs
@@ -3,7 +3,7 @@
     public class ProductPriceUpdate
     {
         public string Id { get; set; }
-        public long NewPriceOpt { get; set; }
-        public long NewPriceRozn { get; set; }
+        public double NewPriceOpt { get; set; }
+        public double NewPriceRozn { get; set; }
     }
 }
